Alpha_V2_Genesis/requirements.txt
flask
flask-cors
yfinance
pandas
numpy
requests
pyngrok
python-dotenv
postgrest
cryptography
langsmith
